DELIA-41736,DELIA-41933 : Buffer based ABR DELIA-41736 : Disable buffer based ABR by default.
Reason for change: Force disabling buffer based ABR for this commit
This will override the App setting also . Added checks for position change for PTS change
Test Procedure: Testcase added in ticket
Risks: Low


DELIA-41736 : Added configurability for low and high buffer value

Reason for change: Added aamp.cfg option for abr low and high buffer value
These values are used for comparing during profile change
Test Procedure: Testcase added in ticket
Risks: Low


DELIA-41736 : Timeout change for segments and playlist.ABR fix

Reason for change: ABR rampdown on delayed download, which was not happening as
cache life was expired and ABR always returns same profile. Changed the timeout for
manifest and playlist . Retry is disabled if buffer duration is less than curltimeout

Test Procedure: Testcase added in ticket
Risks: Medium


DELIA-41933 : Crash on Main Manifest timeout

Reason for change: Crash is happening after main manifest timeout .GetBufferDuration
called with no video instance available .

Test Procedure: Testcase added in ticket
Risks: Medium


DELIA-41736 : Prevent ABR rampdown when transition across discontinuity

Reason for change: Across discontinuity transistion ,only 2 fragments are cached ,
which causes the buffer lower than threshold point to trigger rampdown. This caused
Ad quality to go down.Prevent rampdown across discontinuity on low buffer

Test Procedure: Testcase added in ticket
Risks: Low

+++ jsbindings/jsbindings.cpp
@@ -3084,7 +3084,7 @@ static JSValueRef AAMP_setNetworkTimeout(JSContextRef context, JSObjectRef funct
 	}
 	else
 	{
-		long networkTimeout = (long)JSValueToNumber(context, arguments[0], exception);
+		double networkTimeout = (double)JSValueToNumber(context, arguments[0], exception);
 		pAAMP->_aamp->SetNetworkTimeout(networkTimeout);
 	}
